refactoring: event 모니터링 및 에러 관려 컴포넌트 폴더 이동
Event 모니터링 관련 컴포넌트를 EventMonitor 폴더로 이동
Error 표시 관련 컴포넌트는 errorPage 폴더로 이동
