+++ frontend/host-app/src/components/ErrorPage/ErrorComponent.js
