+++ frontend/host-app/src/components/EventMonitor/SwitchTitle.js
@@ -2,10 +2,10 @@ import React, {useContext} from "react";
 import Switch from "@material-ui/core/Switch";
 import Badge from "@material-ui/core/Badge";
 import {makeStyles} from "@material-ui/core";
-import {socketClient, useSocket} from "../libs/socket.io-Client-wrapper";
-import {HostContext} from "../libs/hostContext";
-import {TitleStyle} from "./StyledComponent/ComponentsStyle.js";
-import TitleBox from "./StyledComponent/TitleBox.js";
+import {socketClient, useSocket} from "../../libs/socket.io-Client-wrapper.js";
+import {HostContext} from "../../libs/hostContext.js";
+import {TitleStyle} from "../StyledComponent/ComponentsStyle.js";
+import TitleBox from "../StyledComponent/TitleBox.js";
 
 const useStyles = makeStyles(theme => ({
 	margin: {
